Only split on rightmost colon in vcftobedpe
Also refactored parsing method out of VCF conversion class so it can be
used in lmerge/lsort.
