+++ tests/util_tests.py
@@ -40,6 +40,16 @@ class InputStreamTest(TestCase):
         stream.close()
         self.assertTrue(stream.handle.closed)
 
+class ParseBndAltStringTest(TestCase):
+    def test_bnd_alt_string(self):
+        self.assertEqual(su.parse_bnd_alt_string('A[1:6['), ('[', '1', 6))
+        self.assertEqual(su.parse_bnd_alt_string('A]1:6]'), (']', '1', 6))
+        self.assertEqual(su.parse_bnd_alt_string(']1:6]A'), (']', '1', 6))
+        self.assertEqual(su.parse_bnd_alt_string(']HLA-DQB1*06:09:01:6]A'), (']', 'HLA-DQB1*06:09:01', 6))
+        with self.assertRaises(AssertionError):
+            su.parse_bnd_alt_string(']1:6[A')
+        with self.assertRaises(AssertionError):
+            su.parse_bnd_alt_string('1')
 
 if __name__ == "__main__":
     main()
